+++ packages/core/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.2.0 - 2019-04-24
+
+#### 🚀 Updates
+
+- **[ButtonOrLink]** Rework rel prop logic (#41) ([aa6b686](https://github.com/airbnb/lunar/commit/aa6b686)), closes [#41](https://github.com/airbnb/lunar/issues/41)
+
+**Note:** Version bump only for package @airbnb/lunar
+
+
+
+
+
+## 1.1.5 - 2019-04-24
+
+#### 🚀 Updates
+
+- **[BorderedList]** Add BorderedList component. (#38) ([3cfc4ea](https://github.com/airbnb/lunar/commit/3cfc4ea)), closes [#38](https://github.com/airbnb/lunar/issues/38)
+
+**Note:** Version bump only for package @airbnb/lunar
+
+
+
+
+
+### 1.1.4 - 2019-04-24
+
+#### 🐞 Fixes
+
+- Remove package dev deps and bump root dev deps. (#44) ([4c26b33](https://github.com/airbnb/lunar/commit/4c26b33)), closes [#44](https://github.com/airbnb/lunar/issues/44)
+
+**Note:** Version bump only for package @airbnb/lunar
+
+
+
+
+
+### 1.1.3 - 2019-04-24
+
+#### 🎨 Styles
+
+- **[Input, Button, MenuItem]** Match padding to design specs (#22) ([6b110db](https://github.com/airbnb/lunar/commit/6b110db)), closes [#22](https://github.com/airbnb/lunar/issues/22)
+
+**Note:** Version bump only for package @airbnb/lunar
+
+
+
+
+
+## 1.1.2 - 2019-04-24
+
+#### 🚀 Updates
+
+- **[Grid]** Add horizontal alignment props. Fix offset % definition. (#35) ([7a81dd6](https://github.com/airbnb/lunar/commit/7a81dd6)), closes [#35](https://github.com/airbnb/lunar/issues/35)
+
+**Note:** Version bump only for package @airbnb/lunar
+
+
+
+
+
+### 1.1.1 - 2019-04-23
+
+#### 🛠 Internals
+
+- Switch to Storybook from Styleguidist (#19) ([129e5c7](https://github.com/airbnb/lunar/commit/129e5c7)), closes [#19](https://github.com/airbnb/lunar/issues/19)
+
+**Note:** Version bump only for package @airbnb/lunar
+
+
+
+
+
 # 1.0.0 - 2019-04-04
 
 #### 🎉 Release