+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 1.0.5 - 2019-04-24
+
+#### 🐞 Fixes
+
+- Remove package dev deps and bump root dev deps. (#44) ([4c26b33](https://github.com/airbnb/lunar/commit/4c26b33)), closes [#44](https://github.com/airbnb/lunar/issues/44)
+
+**Note:** Version bump only for package @airbnb/lunar-forms
+
+
+
+
+
+### 1.0.4 - 2019-04-24
+
+**Note:** Version bump only for package @airbnb/lunar-forms
+
+
+
+
+
+### 1.0.3 - 2019-04-24
+
+**Note:** Version bump only for package @airbnb/lunar-forms
+
+
+
+
+
+### 1.0.2 - 2019-04-23
+
+#### 🛠 Internals
+
+- Switch to Storybook from Styleguidist (#19) ([129e5c7](https://github.com/airbnb/lunar/commit/129e5c7)), closes [#19](https://github.com/airbnb/lunar/issues/19)
+
+**Note:** Version bump only for package @airbnb/lunar-forms
+
+
+
+
+
 # 1.0.0 - 2019-04-04
 
 #### 🎉 Release